Separate log functions into their own file
